fix warning (can’t use std::size_t for signal arguments)

+++ loader.cpp
@@ -278,7 +278,7 @@ void Loader::Worker::markOcCubeAsModified(const CoordOfCube &cubeCoord, const in
     OcModifiedCacheQueue[static_cast<std::size_t>(std::log2(magnification))].emplace(cubeCoord);
 }
 
-void Loader::Worker::snappyCacheSupplySnappy(const CoordOfCube cubeCoord, const int cubeMagnification, const std::string cube) {
+void Loader::Worker::snappyCacheSupplySnappy(const CoordOfCube cubeCoord, const quint64 cubeMagnification, const std::string cube) {
     if (cubeMagnification >= snappyCache.size()) {
         qWarning() << QObject::tr("ignored snappy cube (%1, %2, %3) for higher than available log2(mag) = %4 ≥ %5)")
                       .arg(cubeCoord.x).arg(cubeCoord.y).arg(cubeCoord.z).arg(cubeMagnification).arg(snappyCache.size());
